Repaired Null Value Notification Error.
Set with Null value will get the reference changed to "Not Setup" and also will notify observers, if the previous value was not null.

Additional test case shells have been added to account for missing logic checks. At present, they fail on NotImplementedException.  The code will need to be filled in.